+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changed
 * Change toml parser to tomli
 * Use dj_settings to parse the mappings file
 * Replaced -j/--json and -o/--toml flags with a single -m/--mapping one
+* Require a yashiro section in the config
 
 `0.8.0`_ - 2022-01-10
 ---------------------